devela_micros::_devela

Type Alias OptRes

pub type OptRes<T, E> = Option<Result<T, E>>;

⚖️ An optional result type that combines success, failure, and absence.

📍 code/result::OptRes

Combines Option and Result to provide a structured outcome type with three possible states:

A successful value (Ok)

A failure (Err)

No value (None)

It can simplify value presence management in collections lacking Clone or Default, by using Option::take, and enhance control flow in stateful or asynchronous contexts.

See also: OptResExt, sok and serr.

Examples

use devela::{sok, serr, OptRes};

#[derive(Debug, PartialEq)]
struct V(i32);

fn process_results(results: &mut Vec<OptRes<V, &str>>) {
    println!("Processing...");
    let mut iter = results.iter_mut();
    while let Some(opt_res) = iter.next() {
        if let Some(res) = opt_res.take() {
            match res {
                Ok(mut data) => {
                    println!("  Ok({})", data.0);
                    data.0 += 1; // modify the value
                    *opt_res = sok(data); // and put it back
                }
                Err(err) => {
                    println!("  Err({err})");
                    // leave the current None value
                }
            }
        } else {
            println!("  None");
            *opt_res = serr("Beta"); // replace the None with an error
        }
    }
}

let mut results: Vec<OptRes<V, &str>> = Vec::new();
results.push(None);
results.push(sok(V(10)));
results.push(serr("Alpha"));
results.push(sok(V(20)));

assert_eq![results, vec![None, sok(V(10)), serr("Alpha"), sok(V(20))]];
process_results(&mut results);
assert_eq![results, vec![serr("Beta"), sok(V(11)), None, sok(V(21))]];
process_results(&mut results);
assert_eq![results, vec![None, sok(V(12)), serr("Beta"), sok(V(22))]];

It should print:

Processing...
  None
  Ok(10)
  Err(Alpha)
  Ok(20)
Processing...
  Err(Beta)
  Ok(11)
  None
  Ok(21)

Aliased Type

pub enum OptRes<T, E> {
    None,
    Some(Result<T, E>),
}

Variants

None

No value.

Some(Result<T, E>)

Some value of type T.